Rename numdiff_np.py to derivatives.py.
