+++ docs/source/optimization/general_options.rst
@@ -7,3 +7,7 @@ You can pass a dictionary with general options to the
 :func:`~estimagic.optimization.optimize.minimize` or
 :func:`~estimagic.optimization.optimize.maximize` function. The following options are
 available.
+
+Not that this argument cannot be a list even if several
+
+- "keep_dashboard_alive" (bool): If True, the dashboard is
